multi_open made generic to simplify integration of fflonk without code duplication
